+++ packages/skeleton/src/lib/components/TreeView/types.ts
@@ -1,18 +1,18 @@
+import type { ComponentType } from 'svelte';
+
 export interface TreeViewNode {
-	/** Main content. accepts HTML. */
-	content: string;
-	/** Lead content. accepts HTML. */
-	lead?: string;
-	/** Set open by default on load. */
-	open?: boolean;
-	/** Set the tree disabled state. */
-	disabled?: boolean;
+	/** Nodes Unique ID */
+	id: string;
+	/** Main content. accepts HTML or svelte component. */
+	content: string | ComponentType;
+	/** Main content props. only used when the Content is a svelte component. */
+	contentProps?: object;
+	/** Lead content. accepts HTML or svelte component. */
+	lead?: string | ComponentType;
+	/** lead props. only used when the Lead is a svelte component. */
+	leadProps?: object;
 	/** children nodes. */
 	children?: TreeViewNode[];
 	/** Set the input's value. */
 	value?: unknown;
-	/** input checked */
-	checked?: boolean;
-	/** input is set to indeterminate, only availabe in multiple selection mode. */
-	indeterminate?: boolean;
 }