+++ sites/skeleton.dev/src/routes/(inner)/components/progress-bars/+page.svelte
@@ -41,13 +41,13 @@
 	<svelte:fragment slot="sandbox">
 		<DocsPreview>
 			<svelte:fragment slot="preview">
-				<ProgressBar labelledBy="Progress Bar" bind:value={props.value} max={props.max} />
+				<ProgressBar bind:value={props.value} max={props.max} />
 			</svelte:fragment>
 			<svelte:fragment slot="footer">
 				<div class="w-48 mx-auto"><input type="range" min="0" bind:value={props.value} max={props.max} step={props.step} /></div>
 			</svelte:fragment>
 			<svelte:fragment slot="source">
-				<CodeBlock language="html" code={`<ProgressBar labelledBy="Progress Bar" value={${props.value}} max={${props.max}} />`} />
+				<CodeBlock language="html" code={`<ProgressBar value={${props.value}} max={${props.max}} />`} />
 			</svelte:fragment>
 		</DocsPreview>
 	</svelte:fragment>
@@ -78,6 +78,34 @@
 				<svelte:fragment slot="source">
 					<CodeBlock language="html" code={`<ProgressBar />`} />
 					<CodeBlock language="html" code={`<ProgressBar value={undefined} />`} />
+				</svelte:fragment>
+			</DocsPreview>
+			<h3 class="h3">Custom Animation</h3>
+			<p>Use the <code class="code">animIndeterminate</code> prop to pass a custom CSS animation for the progress bar meter.</p>
+			<DocsPreview background="neutral">
+				<svelte:fragment slot="preview">
+					<div class="w-full space-y-10">
+						<ProgressBar animIndeterminate="anim-progress-bar-example" />
+					</div>
+				</svelte:fragment>
+				<svelte:fragment slot="source">
+					<p>Define your new animation class in your global stylesheet <code class="code">app.postcss</code>.</p>
+					<CodeBlock
+						language="css"
+						code={`
+.anim-progress-bar {
+	transform-origin: 0% 50%;
+	animation: anim-progress-bar 2s infinite linear;
+}
+@keyframes anim-progress-bar {
+	0% { transform: translateX(50%) scaleX(0.5); }
+	50% { transform: translateX(0) scaleX(0.5); }
+	100% { transform: translateX(50%) scaleX(0.5); }
+}
+`}
+					/>
+					<p>Append the animation class to via the <code class="code">animIndeterminate</code> prop.</p>
+					<CodeBlock language="html" code={`<ProgressBar animIndeterminate="anim-progress-bar"/>`} />
 				</svelte:fragment>
 			</DocsPreview>
 		</section>