+++ sites/skeleton.dev/src/routes/(inner)/components/tables/+page.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-	import { faker } from '@faker-js/faker';
 	import { writable, type Writable } from 'svelte/store';
 	// Docs
 	import DocsShell from '$lib/layouts/DocsShell/DocsShell.svelte';
@@ -46,15 +45,19 @@
 	};
 
 	// Local
-	const sourceData = Array(5)
-		.fill(undefined)
-		.map(() => faker.science.chemicalElement());
+	let sourceData = [
+		{ position: 1, name: 'Hydrogen', weight: 1.0079, symbol: 'H' },
+		{ position: 2, name: 'Helium', weight: 4.0026, symbol: 'He' },
+		{ position: 3, name: 'Lithium', weight: 6.941, symbol: 'Li' },
+		{ position: 4, name: 'Beryllium', weight: 9.0122, symbol: 'Be' },
+		{ position: 5, name: 'Boron', weight: 10.811, symbol: 'B' }
+	];
 
 	// Table Simple
 	const tableSimple: TableSource = {
-		head: ['Symbol', 'Name', 'Number'],
-		body: tableMapperValues(sourceData, ['symbol', 'name', 'atomicNumber']),
-		meta: tableMapperValues(sourceData, ['name', 'symbol', 'atomicNumber']),
+		head: ['Symbol', 'Name', 'weight'],
+		body: tableMapperValues(sourceData, ['symbol', 'name', 'weight']),
+		meta: tableMapperValues(sourceData, ['name', 'symbol', 'weight']),
 		foot: ['Total Elements', '', `<span class="badge variant-soft-primary">${sourceData.length} Elements</span>`]
 	};
 
@@ -115,6 +118,7 @@ const tableSimple: TableSource = {
 
 	<!-- Slot: Usage -->
 	<svelte:fragment slot="usage">
+		<!-- Interactive Mode -->
 		<section class="space-y-4">
 			<h2 class="h2">Interactive Mode</h2>
 			<p>
@@ -124,6 +128,29 @@ const tableSimple: TableSource = {
 				<code class="code">meta</code> value.
 			</p>
 			<CodeBlock language="html" code={`<Table ... interactive={true} on:selected={mySelectionHandler} />`} />
+		</section>
+		<!-- Handling Reactive Data -->
+		<section class="space-y-4">
+			<h2 class="h2">Handling Reactive Data</h2>
+			<!-- prettier-ignore -->
+			<p>
+				In some scenarios you may need need to reactive manner. Since <code class="code">TableSource</code> is a Javascript object, it will need to follow Svelte's rules for <a class="anchor" href="https://learn.svelte.dev/tutorial/updating-arrays-and-objects" target="_blank" rel="noreferrer">reactive object updates</a>.
+			</p>
+			<CodeBlock
+				language="ts"
+				code={`
+function setTableSource(): TableSource {
+	return {
+		head: ['Symbol', 'Name', 'weight'],
+		body: tableMapperValues(sourceData, ['symbol', 'name', 'weight']),
+		meta: tableMapperValues(sourceData, ['name', 'symbol', 'weight']),
+		foot: ['Total Elements', '', \`<span class="badge variant-soft-primary">${sourceData.length} Elements</span>\`]
+	};
+}\n
+// If sourceData updates, set the new TableSource values
+$: tableSimple = sourceData ? setTableSource() : undefined;
+		`}
+			/>
 		</section>
 		<!-- Table Utilities -->
 		<section class="space-y-4">